refactor: extract TaskGroupingTreeStorage to a separate file
